requirements.txt
langchain
chromadb
urllib3
PyMuPDF
python-dotenv
unstructured
extract-msg
tabulate
pandoc
pypandoc
tqdm
sentence_transformers
langchain-community 
tiktoken 
langchain-openai 
langchainhub 
langchain-cohere
pymupdf
streamlit
chroma-migrate
langchain 
llama-index 
langchain-experimental 
ollama